tuning for turns year/week

+++ cw-main/src/main/java/com/cwsni/world/model/player/PCountry.java
@@ -189,7 +189,7 @@ public class PCountry implements IPCountry {
 	}
 
 	@Override
-	public long getCasualties() {
+	public double getCasualties() {
 		return country.getCasualties();
 	}
